+++ resources/views/livewire/settings/profile.blade.php
@@ -121,7 +121,10 @@ new class extends Component {
 
             <div class="flex items-center gap-4">
                 <div class="flex items-center justify-end">
-                    <x-daisyui.button variant="primary" type="submit" class="w-full">{{ __('Save') }}</x-daisyui.button>
+                    <x-daisyui.button variant="primary" type="submit" class="w-full">
+                        <x-daisyui.loading-spinner/>
+                        {{ __('Save') }}
+                    </x-daisyui.button>
                 </div>
 
                 <x-action-message class="me-3" on="profile-updated">