+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 2025-03-07
 
 -   Added `x-daisyui.modal` and related components
+-   Added `x-daisyui.loading-spinner` component
 
 ## 2025-03-06